+++ modules/abstract-eth/test/unit/index.ts
@@ -1,0 +1,4 @@
+export * from './transactionBuilder';
+export * from './token';
+export * from './transaction';
+export * from './coin';