+++ modules/sdk-coin-tia/test/unit/tia.ts
@@ -24,6 +24,10 @@ import {
 import should = require('should');
 import nock = require('nock');
 
+import { loadWebAssembly } from '@bitgo/sdk-opensslbytes';
+
+const openSSLBytes = loadWebAssembly().buffer;
+
 describe('TIA', function () {
   let bitgo: TestBitGoAPI;
   let basecoin;
@@ -475,13 +479,16 @@ describe('TIA', function () {
     });
 
     it('should recover funds for non-bitgo recoveries', async function () {
-      const res = await basecoin.recover({
-        userKey: wrwUser.userPrivateKey,
-        backupKey: wrwUser.backupPrivateKey,
-        bitgoKey: wrwUser.bitgoPublicKey,
-        walletPassphrase: wrwUser.walletPassphrase,
-        recoveryDestination: destinationAddress,
-      });
+      const res = await basecoin.recover(
+        {
+          userKey: wrwUser.userPrivateKey,
+          backupKey: wrwUser.backupPrivateKey,
+          bitgoKey: wrwUser.bitgoPublicKey,
+          walletPassphrase: wrwUser.walletPassphrase,
+          recoveryDestination: destinationAddress,
+        },
+        openSSLBytes
+      );
       res.should.not.be.empty();
       res.should.hasOwnProperty('serializedTx');
       sandBox.assert.calledOnce(basecoin.getAccountBalance);
@@ -527,46 +534,58 @@ describe('TIA', function () {
 
     it('should throw error if backupkey is not present', async function () {
       await basecoin
-        .recover({
-          userKey: wrwUser.userPrivateKey,
-          bitgoKey: wrwUser.bitgoPublicKey,
-          walletPassphrase: wrwUser.walletPassphrase,
-          recoveryDestination: destinationAddress,
-        })
+        .recover(
+          {
+            userKey: wrwUser.userPrivateKey,
+            bitgoKey: wrwUser.bitgoPublicKey,
+            walletPassphrase: wrwUser.walletPassphrase,
+            recoveryDestination: destinationAddress,
+          },
+          openSSLBytes
+        )
         .should.rejectedWith('missing backupKey');
     });
 
     it('should throw error if userkey is not present', async function () {
       await basecoin
-        .recover({
-          backupKey: wrwUser.backupPrivateKey,
-          bitgoKey: wrwUser.bitgoPublicKey,
-          walletPassphrase: wrwUser.walletPassphrase,
-          recoveryDestination: destinationAddress,
-        })
+        .recover(
+          {
+            backupKey: wrwUser.backupPrivateKey,
+            bitgoKey: wrwUser.bitgoPublicKey,
+            walletPassphrase: wrwUser.walletPassphrase,
+            recoveryDestination: destinationAddress,
+          },
+          openSSLBytes
+        )
         .should.rejectedWith('missing userKey');
     });
 
     it('should throw error if wallet passphrase is not present', async function () {
       await basecoin
-        .recover({
-          userKey: wrwUser.userPrivateKey,
-          backupKey: wrwUser.backupPrivateKey,
-          bitgoKey: wrwUser.bitgoPublicKey,
-          recoveryDestination: destinationAddress,
-        })
+        .recover(
+          {
+            userKey: wrwUser.userPrivateKey,
+            backupKey: wrwUser.backupPrivateKey,
+            bitgoKey: wrwUser.bitgoPublicKey,
+            recoveryDestination: destinationAddress,
+          },
+          openSSLBytes
+        )
         .should.rejectedWith('missing wallet passphrase');
     });
 
     it('should throw error if there is no balance', async function () {
       await basecoin
-        .recover({
-          userKey: wrwUser.userPrivateKey,
-          backupKey: wrwUser.backupPrivateKey,
-          bitgoKey: wrwUser.bitgoPublicKey,
-          walletPassphrase: wrwUser.walletPassphrase,
-          recoveryDestination: destinationAddress,
-        })
+        .recover(
+          {
+            userKey: wrwUser.userPrivateKey,
+            backupKey: wrwUser.backupPrivateKey,
+            bitgoKey: wrwUser.bitgoPublicKey,
+            walletPassphrase: wrwUser.walletPassphrase,
+            recoveryDestination: destinationAddress,
+          },
+          openSSLBytes
+        )
         .should.rejectedWith('Did not have enough funds to recover');
     });
   });