+++ modules/sdk-lib-mpc/src/safePrime.ts
@@ -1,14 +1,14 @@
 import { OpenSSL } from './openssl';
 
-export async function generateSafePrime(bitlength: number): Promise<bigint> {
+export async function generateSafePrime(bitlength: number, openSSLBytes: Uint8Array): Promise<bigint> {
   const openSSL = new OpenSSL();
-  await openSSL.init();
+  await openSSL.init(openSSLBytes);
   return openSSL.generateSafePrime(bitlength);
 }
 
-export async function generateSafePrimes(bitLengths: number[]): Promise<bigint[]> {
+export async function generateSafePrimes(bitLengths: number[], openSSLBytes: Uint8Array): Promise<bigint[]> {
   const openSSL = new OpenSSL();
-  await openSSL.init();
+  await openSSL.init(openSSLBytes);
   const promises: Promise<bigint>[] = bitLengths.map((bitlength: number) => {
     return openSSL.generateSafePrime(bitlength);
   });