+++ modules/abstract-cosmos/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [10.1.1](https://github.com/BitGo/BitGoJS/compare/@bitgo/abstract-cosmos@10.1.0...@bitgo/abstract-cosmos@10.1.1) (2024-08-13)
+
+**Note:** Version bump only for package @bitgo/abstract-cosmos
+
 # [10.1.0](https://github.com/BitGo/BitGoJS/compare/@bitgo/abstract-cosmos@10.0.28...@bitgo/abstract-cosmos@10.1.0) (2024-08-07)
 
 ### Features