+++ modules/sdk-coin-atom/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [12.1.1](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-atom@12.1.0...@bitgo/sdk-coin-atom@12.1.1) (2024-08-13)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-atom
+
 # [12.1.0](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-atom@12.0.28...@bitgo/sdk-coin-atom@12.1.0) (2024-08-07)
 
 ### Features