+++ modules/sdk-coin-sui/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.4.1](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-sui@5.4.0...@bitgo/sdk-coin-sui@5.4.1) (2024-08-13)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-sui
+
 # [5.4.0](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-sui@5.3.1...@bitgo/sdk-coin-sui@5.4.0) (2024-08-07)
 
 ### Bug Fixes