+++ modules/sdk-coin-xtz/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.28](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-xtz@2.1.27...@bitgo/sdk-coin-xtz@2.1.28) (2024-08-13)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-xtz
+
 ## [2.1.27](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-xtz@2.1.26...@bitgo/sdk-coin-xtz@2.1.27) (2024-08-07)
 
 **Note:** Version bump only for package @bitgo/sdk-coin-xtz